reboot/requirements.txt
appdirs==1.4.3
astroid==2.4.2
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
defusedxml==0.6.0
diff-match-patch==20181111
distlib==0.3.0
distro==1.5.0
dj-database-url==0.5.0
Django==2.1.5
django-bootstrap-datepicker-plus==3.0.5
django-cleanup==5.0.0
django-heroku==0.3.1
django-import-export==2.1.0
djangorestframework==3.11.1
et-xmlfile==1.0.1
gunicorn==20.0.4
html5lib==1.0.1
idna==2.9
ipaddr==2.2.0
isort==4.3.21
jdcal==1.4.1
lazy-object-proxy==1.4.3
lockfile==0.12.2
MarkupPy==1.14
mccabe==0.6.1
msgpack==0.6.2
odfpy==1.4.1
openpyxl==3.0.3
packaging==20.3
pep517==0.8.2
Pillow==7.1.2
progress==1.5
psycopg2==2.8.5
psycopg2-binary==2.8.5
pylint==2.5.3
pyparsing==2.4.7
pytoml==0.1.21
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
retrying==1.3.3
six==1.14.0
tablib==2.0.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.8
webencodings==0.5.1
whitenoise==5.1.0
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
